+++ libosmscout-import/include/osmscout/import/GenRelAreaDat.h
@@ -153,7 +153,7 @@ namespace osmscout {
                      std::list<MultipolygonPart>& groups,
                      GroupingState& state,
                      size_t topIndex,
-                     size_t id);
+                     uint8_t id);
 
     bool BuildRings(const TypeConfig& typeConfig,
                     const ImportParameter& parameter,
